+++ tests/po_lib/__init__.py
@@ -12,8 +12,9 @@ from .. import po_lib_sub  # noqa: F401
 
 
 class POBase(ItemPage):
-    expected_overrides: Type[ItemPage]
+    expected_instead_of: Type[ItemPage]
     expected_patterns: Patterns
+    expected_to_return: Any = None
     expected_meta: Dict[str, Any]
 
 
@@ -25,19 +26,22 @@ class POTopLevelOverriden2(ItemPage):
     ...
 
 
-# This first annotation is ignored. A single annotation per registry is allowed
-@handle_urls("example.com", overrides=POTopLevelOverriden1)
+# This first decorator is ignored. A single ``ApplyRule`` with the same Page
+# Object to be used per registry is allowed.
+@handle_urls("example.com", instead_of=POTopLevelOverriden1)
 @handle_urls(
-    "example.com", overrides=POTopLevelOverriden1, exclude="/*.jpg|", priority=300
+    "example.com", instead_of=POTopLevelOverriden1, exclude="/*.jpg|", priority=300
 )
 class POTopLevel1(POBase):
-    expected_overrides = POTopLevelOverriden1
+    expected_instead_of = POTopLevelOverriden1
     expected_patterns = Patterns(["example.com"], ["/*.jpg|"], priority=300)
+    expected_to_return = None
     expected_meta = {}  # type: ignore
 
 
-@handle_urls("example.com", overrides=POTopLevelOverriden2)
+@handle_urls("example.com", instead_of=POTopLevelOverriden2)
 class POTopLevel2(POBase):
-    expected_overrides = POTopLevelOverriden2
+    expected_instead_of = POTopLevelOverriden2
     expected_patterns = Patterns(["example.com"])
+    expected_to_return = None
     expected_meta = {}  # type: ignore