+++ tests/po_lib/nested_package/a_nested_module.py
@@ -11,11 +11,12 @@ class PONestedModuleOverriden(ItemPage):
 @handle_urls(
     include=["example.com", "example.org"],
     exclude=["/*.jpg|"],
-    overrides=PONestedModuleOverriden,
+    instead_of=PONestedModuleOverriden,
 )
 class PONestedModule(POBase):
-    expected_overrides = PONestedModuleOverriden
+    expected_instead_of = PONestedModuleOverriden
     expected_patterns = Patterns(
         include=["example.com", "example.org"], exclude=["/*.jpg|"]
     )
+    expected_to_return = None
     expected_meta = {}  # type: ignore