+++ web_poet/_typing.py
@@ -18,7 +18,14 @@ def is_generic_alias(obj) -> bool:
 
 
 def get_generic_parameter(cls):
-    for base in cls.__orig_bases__:
+    for base in getattr(cls, "__orig_bases__", []):
         if is_generic_alias(base):
             args = _get_args(base)
             return args[0]
+
+
+def get_item_cls(cls, default=None):
+    param = get_generic_parameter(cls)
+    if param is None or isinstance(param, typing.TypeVar):  # class is not parametrized
+        return default
+    return param